requirements.txt
colorgram.py==1.2.0
prettytable==3.10.0
ruff==0.5.6
pandas==2.2.3
pyperclip==1.9.0
python-dotenv==1.0.1
